requirements.txt
blinker==1.6.2
click==8.1.3
dnspython==2.7.0
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.10.1
python-dotenv==1.0.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==2.3.3
marshmallow==3.23.0
Flask-RESTful==0.3.10